refactor: useFetchData 훅 생성 후 분리
